requirements.txt
fastapi==0.85.0
h5py==3.7.0
keras==2.10.0
Keras-Preprocessing==1.1.2
matplotlib==3.6.0
numpy==1.23.3
Pillow==9.2.0
tensorflow-cpu
uvicorn==0.18.3
gunicorn==20.0.4
python-multipart==0.0.5